Check against "python" starting python3
Arch Linux reportedly changed their default setup so that plain
"python" starts python3. They only have a version compatible with
Python 2 available under the name "python2". The scripts require
Python version 2, and I don't want to change them to 3 quite yet. This
means there's no good way to specify the interpreter that'd work both
on Arch and on a normal setup (on Arch you'd need to use "python2",
but that name is not available on Debian for example).

Abort with an explicit error message about version incompatibility if
script/helpers.py is imported under Python 3 (before the result was an
obscure error). This should make it more obvious for Arch Linux users
that it's their distro causing problems and how to work around it.