Switch ffmpeg-mt submodule to use Libav
FFmpeg-mt functionality has been merged to Libav and the original
FFmpeg-mt tree is now obsolete. Switch the submodule to a Libav-based
branch. Currently the only difference from upstream Libav is that
GCC flags use -O2 instead of -O3. I'm leaving possible changes to
the names used for the submodule and other things for later.