script/helpers.py: remove stray newline from git parameter
A URL passed to "git clone" contained a newline character at the end
left over from the output of the "git config --get" call used to query
it. Apparently the newline was ignored by git and did not trigger any
problems. Add a .strip() to remove it in any case.