Update libav submodule (libpostproc dropped)
This version of Libav drops libpostproc, which is now maintained as a
separate project. I'm not adding an internal libpostproc to this
repository for now at least. The practical effect is that the player
will only support the "pp" video filter if a system installation of
libpostproc exists.

Also, older system installations of still-under-Libav libpostproc are
likely to fail: versions from Libav declared a dependency on an
_exact_ libavutil version number, which would cause "pkg-config
--cflags libpostproc" to fail with an error message like "Package
'libpostproc' requires 'libavutil = 51.21.0' but version of libavutil
is 51.24.0" when libavutil comes from our newer internal version.