Revert back to previous libav version
The commit fixing h264_vdpau crashes when called with 10-bit input
instead causes a crash when using H.264 parsing (used at least when
demuxing H.264 in mpeg containers or AVI). Change back to the previous
version; the crash with h264_vdpau is a less serious issue as hardware
decoding wouldn't actually succeed with 10-bit input anyway and you
can manually switch to software decoding to work around the problem.

The problem causing the crash is that the parser calls some H.264
decoder functions with codec pointer set to NULL, while the vdpau fix
tries to check codec data to see whether it's the h264_vdpau decoder
or the software one.